requirements.txt
certifi==2016.2.28
Django==2.0.1
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
Serializer==0.2.1
simplejson==3.13.2
wincertstore==0.2
